+++ lib/image_processing/vips.rb
@@ -49,6 +49,14 @@ module ImageProcessing
         image.gravity(gravity, width, height, **embed_options)
       end
 
+      def rotate(image, degrees, **options)
+        if degrees % 90 == 0
+          image.rot(:"d#{degrees % 360}")
+        else
+          image.similarity(angle: degrees, **options)
+        end
+      end
+
       def load_image(path_or_image, autorot: true, **options)
         if path_or_image.is_a?(::Vips::Image)
           image = path_or_image