+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* [minimagick, vips] Add `#rotate` function (@janko-m)
+
 ## 1.2.0 (2018-04-18)
 
 * [minimagick] Allow appending "+" operators in `#loader` and `#saver` using the value `false` (@janko-m)